Evaluations might not be done by you
Evaluation of software might not be done by you
(especially if you're an end-user of software and don't know
how to develop software). You might hire someone else, or get
others' reviews. Broaden the last point to note that the need
is for evaluation, it doesn't have to be done by the reader.

I also added a link to security-reviews, as part of the way to find
evaluations by others.

While I'm at it, I fixed invalid UTF-8 characters in point 6, which
probably should be a separate pull request but I'm in a hurry :-).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>