Add sample hyperlink for -fstack-protector-strong
It would be nice if readers could click on a compiler flag in a
summary and immediately see the details about that flag.

I propose adding hyperlinks to implement that. It would make the markdown
a *little* messier, but not outrageously so, and it would make the
document *much* easier to navigate. I think we need to make it *easy*
for readers to get from a summary to the details they need, or they
won't read the details they need.

This adds an internal hyperlink from a flag description in a summary
to an anchor on the details. The anchor is a span with id; I could have
used a name="..." instead, but many seem to be preferring id so I used
that. The id is exactly the flag name (e.g., `-fstack-protector-strong`);
leading dashes *are* allowed in ids.

NB: The sample entry makes 1 table entry longer than the others; if they
*all* do that, then columns could be re-aligned again without lots of
space.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>