Update Compiler-Options-Hardening-Guide-for-C-and-C++.md
Make it clear the compiler version column is the first version of the compiler offering this flag and not the only one.

Honestly I created this page:
[^CMU2018]: Carnegie Mellon University (CMU), [Top 10 Secure Coding Practices](https://wiki.sei.cmu.edu/confluence/display/seccode/Top+10+Secure+Coding+Practices)
For people who wouldn't stop asking me for "something easy".

I do not recommend `-Werror` as I think it creates more problems than it is worth, although I do agree all warnings should be addressed.  Minimally I wouldn't include this because there is no clear consensus this option is beneficial.  You could mention it and what it does, but I would stay neutral WRT to recommending it.

Signed-off-by: Robert C. Seacord <rcseacord@gmail.com>
Signed-off-by: noamd <noam@legitsecurity.com>