Add an index.md to /docs of "released work"
Add an "index.md" file so that people can view
<https://best.openssf.org> and see all the *released* work
from the OpenSSF Best Practices Working Group.

The goal is to make this more discoverable.

We *separately* need to create a "landing page" for developers,
one that would cover all OpenSSF work (at least).
That is a *separate* idea, I currently expect that to
be developed in `./developer.md`. But let's quickly get
this "general list" added, so it'll be easy to find this info!

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>
Signed-off-by: noamd <noam@legitsecurity.com>