Omit leading slash in _config.yml path
It turns out that the _config.yml path value
doesn't use a leading "/" when referring to files.
Remove it. See:
https://jekyllrb.com/docs/configuration/front-matter-defaults/

This will let us set the descriptions of individual files
*without* adding a yml header to the markdown itself.
Currently people are jumping directly to the markdown,
so we want the markdown as readable as we can make it until
people generally use the publishing website best.opennsf.org.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>
Signed-off-by: noamd <noam@legitsecurity.com>