Initial cut at a "developer landing page"
I'm sure I'm missing thing and much needs to be improved,
but this is a first cut at creating a "landing page" for
developers.

I tried to use the "best links available" but some resources
don't have a helpful and easy "here's how to get started using this"
page. So I did what I could, and this should spur on fixing that.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>