Remove "critical" for sos.dev
The word "critical" is used in different ways, and people found
it confusing when talking about sos.dev. The simple solution
is to simply remove the confusing word.

We'll instead depend on our link to the sos.dev FAQ.
That FAQ gives a better explanation than we can manage in just
a few words.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>