Add dates to concise guides
This adds YYYY-MM-DD dates to:

* `Concise-Guide-for-Developing-More-Secure-Software.md`
* `Concise-Guide-for-Evaluating-Open-Source-Software.md`

The TAC approved WGs, projects, and SIGs using this
"Simplest Possible Process" for publication if they wanted to use it in
<https://github.com/ossf/tac/issues/176>

However, it was noted that dates were really important to include.
There are complex ways to handle this, but they turn out to be
complicated to do automatically, while it's quite simple to simply
insert the date in the document itself - so let's do that.

This adds the date just under the title, with the (WG) author name.
That makes it very easy to find and other documents also do this.

I'm using ISO 8601 format, YYYY-MM-DD. This is easily understood
by all (the "traditional" formats of the US and Europe swap the order
of months and days, so that's best avoided), it uses less space than
most other formats, and it sorts well.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>