Switch to font IBMPlexSans for headers
Many found the font Teko to be too hard to read, so
the OpenSSF brand guidelines at
<https://openssf.org/brand-guidelines/>
were updated to switch from Teko to IBM Plex Sans.

This commit switches to the font IBM Plex Sans.

Note that we'll continue to use Cairo for body text
(people seem fine with that).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>