Fix markdownlint errors due to indentation
As noted in
[CommonMark list-items](https://spec.commonmark.org/0.30/#list-items),
"The most important thing to notice is that the position of the text
after the list marker determines how much indentation is needed in
subsequent blocks in the list item. If the list marker takes up two
spaces of indentation, and there are three spaces between the list
marker and the next character other than a space or tab, then blocks
must be indented five spaces in order to fall under the list item."

This means that renumbering an item may require indentation changes
to match.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>