best -> good
I don't think we can easily argue that this particular configuration
is the "best balance" in all cases. As Disconnect3d from Trail of Bits
noted in Slack, "citation needed :-)". Trade-offs are *way* too
complicated to baldly state something is always the "best balance".

So instead, just say it's a "good balance". There may be a better
balance for your circumstance, but since this is a good balance, we
can recommend it.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>