Precompute correct regex once
The previous code was very inefficient, it acquired and
compiled the expression for the correct answer on every
change or key-up. Let's compute the correct compiled version
once, on startup, so users never experience a slowdown.

It's likely we will need to support multiple input values
instead of just one. This optimization will be even more
important if that happens, so we may as well start now.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>