Fix tooltips for buttons
If a button doesn't define a title it doesn't necessarily result
in the value being "undefined" or "null" - empty string is possible.
I find that odd, but it's reality.

This code changes so that if a button title is falsy at all
(null, undefined, or empty string), we set its value.
That way every button has a tooltip.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>