Show individual statuses as themselves (bugfix)
Previously if there were multiple lab entrants, their
background color would be set based on cumulative correctness.
This was a bug; they should be shown separately.

The only way to notice this is to fill in correct answers
in the wrong order, so it's not obvious, but we want to handle
this unusual case correctly.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>