Checker: Generalize labs to work per-form on reset
This commit begins generalizing the labs so that there
can be more than one form in an HTML page.
This means that all buttons need a special class, instead of
a unique id, and that the code needs to use that class instead.

This commit modifies checker.js to use these new classes,
and reset now only resets its specific form.
The hint and giveUp buttons are still global, but this adds the
infrastructure for changing that.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>