Change lab html style to refer to best.openssf.org
Modify the lab HTML files to always try to load from
best.openssf.org. This is no change for accesses to best.openssf.org,
but it means that if you're using the lab locally *and* have
internet access, you'll see the same thing. This is an improvement
for local testing and individualized use.

Neither approach will allow the use of styles for disconnected operations.
That said, the labs will still *work* they'll just look different.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>