Tweak tone lab README
The README has helpful information, but it's really a roadmap
and people at first didn't realize that. Modify the tone so
it's clearly a roadmap, show stages, and make it clear that these
are "planned" not merely "needed".

This commit is per review/comment during an OpenSSF education SIG
meeting.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>