Provide recommendation to counter xz utils style attack
The malicious attack on the xz utils slipped through many
defenses because the "source" package included pre-generated
malicious code. This meant that review of the source code
(e.g., as seen by git) couldn't find the problem.

This proposes a best practices to counter it. The text is longer
than I'd like, but it's hard to make it short, and this was a
worrying attack so I think it's reasonable to say this.

We'll probably need to renumber this proposal if we also add
the proposed text to counter attacks like polyfill.io:
https://github.com/ossf/wg-best-practices-os-developers/pull/559
... but I think that's okay!

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>