shell-injection.html: Improve hints further
Raw strings are recommended but not really *required* in this case,
so recommend them in hints, but don't *mandate* them in the answer.
Allow r'...' and r"..." (both are allowed in Python).

Be pickier about whitespace in a line but NOT in parentheses - in Python
a newline ENDS a statement if it's not bracketed.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>