shell-argument-injection.html: Tweak allowed answer
Tweak required answer to be easier to read *and* more generous.
JavaScript allows string constants with '...' or "..." or `...`
(the last are a template, but you don't HAVE to use {...} inside them).
The shell: false parameter isn't required.

Also, whitespace is interpreted as "0 or more whitespace" in a pattern;
use that to make the patterns easier to read.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>