Python guide: template for contributing to the project
The change aims to create a simple template for creating README files
for Secure Coding Python Guides. By using the template, we can ensure
that all of the rule descriptions follow the same format in regards to
page sections, code examples, bibliography, etc. The template also
contains the link to quoting reference guide we have decided to follow.

The template should be mentioned on the landing page referenced in #520

Signed-off-by: edanhub <hubert.daniszewski@ericsson.com>