Improve lab testing (#673)
* Improve lab testing

Improve how labs are tested.

Change how the "examples" are used in the hint section
so that people provide full examples up to the tested index,
and *document* this (before it wasn't clear how do this).
By using this approach we will *consistently* test hints.

In the longer term, we might remove the constraint on which
hints are returned during the testing, but that requires
determining which forms cover which indexes, which is more
complex. By asking for the fields now, we make it easier
to make that change later.

This makes changes in labs to *use* the new testing mechanism
in hint examples.

This also adds the script "mass-test", which automatically
tests all labs. It does this by opening them all in the web browser,
triggering the self-test in each lab. You have to manually look
at each page, which is a little annoying. That said,
that takes very little time. Automatically opening
every page is a big improvement in terms of automated testing.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Remove trailing space

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

---------

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>