Provide recommendation to counter xz utils style attack (#560)
* Provide recommendation to counter xz utils style attack

The malicious attack on the xz utils slipped through many
defenses because the "source" package included pre-generated
malicious code. This meant that review of the source code
(e.g., as seen by git) couldn't find the problem.

This proposes a best practices to counter it. The text is longer
than I'd like, but it's hard to make it short, and this was a
worrying attack so I think it's reasonable to say this.

We'll probably need to renumber this proposal if we also add
the proposed text to counter attacks like polyfill.io:
https://github.com/ossf/wg-best-practices-os-developers/pull/559
... but I think that's okay!

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Fix grammar nit in xz utils response

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Clarify text

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Fix emphasis for markdownlint

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Update docs/Concise-Guide-for-Developing-More-Secure-Software.md

Co-authored-by: j-k <jack@control-plane.io>
Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Clarify source package should only include VCS materials

The source package should be a copy or subset of the VCS materials.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Update docs/Concise-Guide-for-Developing-More-Secure-Software.md

Co-authored-by: Jordan Harband <ljharb@gmail.com>
Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

---------

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>
Co-authored-by: j-k <jack@control-plane.io>
Co-authored-by: Jordan Harband <ljharb@gmail.com>