Create/implement separate hello.js for hello.html config
Currently configuration is mostly in a YAML file embedded in the
lab's HTML. This means that the lab configuration cannot be shared
between locales (translations), which is annoying.

We can't easily load YAML or JSON from a separate file when
running or testing locally. However, we *can* easily load a
separate file if it's in JavaScript format.

This switches the configuration of lab hello.html so it uses the
configuration in hello.js.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>