Labs: Make "give up" require a certain number of seconds
To encourage people to *try* to do the labs, if they pick
"give up" too soon after loading, we'll instead tell them to
try harder first. It also shows the number of seconds elapsed
(a strong hint that this is timed).

This intentionally does not show the minimum time. The purpose is
to get people to work on the lab, not to encourage them to
game the timer.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>