Fix typo in return value from README (fixes #652) (#716)
* Fix typo in return value from README (fixes #652)

The code snippet incorrectly returns `res` when it should be returning `ret`. This doesn't really have an effect, since the sample demonstrates failing early.

Signed-off-by: Michael Scovetta <michael.scovetta@gmail.com>

* Fix typo in return value from README (fixes #652)

The same code is in both the README and compliant03.py. This change fixes the typo in compliant03.py.

Checked by Hubert who was working on these code examples

Signed-off-by: Michael Scovetta <michael.scovetta@gmail.com>

---------

Signed-off-by: Michael Scovetta <michael.scovetta@gmail.com>