Create/implement separate hello.js for hello.html config (#728)
* Create/implement separate hello.js for hello.html config

Currently configuration is mostly in a YAML file embedded in the
lab's HTML. This means that the lab configuration cannot be shared
between locales (translations), which is annoying.

We can't easily load YAML or JSON from a separate file when
running or testing locally. However, we *can* easily load a
separate file if it's in JavaScript format.

This switches the configuration of lab hello.html so it uses the
configuration in hello.js.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Simplify hello.js

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* hello.js: Provide Japanese translation examples

This does not *implement* translations, but at least this
includes some translations (which we hope to use in a later
change).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Auto-report info and info2 differences

To make it easier to move from YAML, automatically detect
differences between info and info2. We can create a JavaScript
file that looks clean but is *supposed* to be identical, and
once done, eliminate the old version. This will help ensure
we don't make mistakes during the transition.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Note in markdown the beginning of transition from YAML

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

---------

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>