Handle in-browser translations more gracefully
When Chrome is presented with a natural language, it
can offer to translate the HTML page. Unfortunately,
this caused labs to fail.

This commit modifies to code to work around this, so
that after translation the lab continues to work.
The explanation about the lab is translated.
Unfortunately, this does *not* translate the responses
shown via alert boxes (e.g., hints and congrats).
However, since the lab now *works*, users can copy that text
into something to do a translation.

This commit makes the labs much more accessible to those
who speak natural languages we don't currently have
translations for.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>