Remove YAML processing library
We've switched from using YAML to using JavaScript directly
for storing data. One reason to do that was simply that
it's the only easy way to load data when running locally.

A second reason, however, is that this means we can
completely eliminate the use of this library at run-time.
Not using a library at *all* eliminates the need to maintain it,
update it, etc. It also speeds loading; nothing loads faster
than code you don't load at all :-).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>