Create lab fr_hello.html (French translation)
This is merely a machine language translation, so I'm
sure it's imperfect. However, it's useful for testing our
internationalization support. For example, Firefox has
built-in translation support for French but not yet for
Japanese, and this is a useful sample for testing it.
I can also read some French, so I can more confidently test it.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>