Refactor format() so it's not a global String property
While JavaScript *allows* global properties on Strings to
be added, it's not necessarily a good idea, as those are global
(and could interfere with other things).

To reduce the risk of future problems, use a created function instead,
to avoid conflicts.

This also adds a myAssert and improves internal comments about
indexing as related changes.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>