Make it clear team reviews *translations*
Anyone, including trusted translation team members, is welcome to
note any issues about any PR.

However, in this text, make it clear that the trusted translation team
members are focusing on the *translation*. They might not notice
other issues, nor is it reasonable to expect them to do so.
Let's make sure we're *not* asking them to do anything unreasonable.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>