Be silent on being manufacturer + steward for same bits
Deal with a difference in interpretation by *not*
stating a claim one way or the other.

@maertsen believes that it's possible to be both a manufacturer
and a steward for the same sequence of bits but in different
contexts. Others don't agree with this interpretation.

We don't have the authority to interpret the CRA.
So let's resolve this by saying what *is* clearly known,
citing authority (in this case the EC), and stay silent where it's
not clear what the interpretation is.

If there's a formal interpretation stated later, we can change this
text to reflect that.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>