CWE-1095: show that list modification is dangerous (#940)
Before this commit the wording was that modifying list works but
is not recommended.

But it works as long as no two consecutive elements are deleted,
otherwise part of elements is not checked at all without any
exceptions raised.

Changed README.md, compliant01.py and noncompliant01.py to
demonstrate that.

Signed-off-by: Kyrylo Yatsenko <hedrok@gmail.com>
Signed-off-by: Helge Wehder <myteron@gmail.com>