Tweak intro text of list
The criteria used to be expressed in the form of questions;
now they are identified as concerns and information that can
help you determine how concerned you should be.
However, we didn't update the introductory sentence to match.

This updates the introductory text so that text matches what follows it.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>